+++ CHANGELOG.rst
@@ -37,6 +37,8 @@ Change Log
 - [ADDED] the implementation of `+` operator for action (based on previously available `+=`)
 - [ADDED] A more detailed documentation on the representation of the topology and how to create a backend
 - [ADDED] A easier way to set up the topology in backend (eg. `get_loads_bus`)
+- [ADDED] A easier way to set up the backend, with automatic computation of some attributes (eg. `*_to_sub_pos`,
+  `sub_info`, `dim_topo`) if needed.
 - [IMPROVED] `PandaPowerBackend` now should take less time to when `reset`.
 - [IMPROVED] some speed up in the grid2op computation